pyproject.toml
[project]
name = "errorworks"
version = "0.1.0"
description = "Composable chaos-testing services for various pipelines"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "John Morrissey", email = "john.morrissey@dta.gov.au" }]
maintainers = [{ name = "Digital Transformation Agency" }]
keywords = ["chaos", "testing", "llm", "web", "fault-injection", "resilience"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    # Configuration
    "pydantic>=2.12,<3",
    "pyyaml>=6.0,<7",

    # CLI
    "typer>=0.21,<1",

    # Templating (response/content generation)
    "jinja2>=3.1,<4",

    # Logging
    "structlog>=25.0,<26",

    # HTTP server
    "starlette>=0.45,<1",
    "uvicorn>=0.34,<1",

    # HTTP client (for test fixtures)
    "httpx>=0.27,<1",

    # MCP server for ChaosLLM metrics analysis
    "mcp>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0,<10",
    "pytest-cov>=4.1,<6",
    "pytest-asyncio>=1.0,<2",
    "pytest-timeout>=2.3,<3",
    "hypothesis>=6.98,<7",
    "mypy>=1.19,<2",
    "ruff>=0.15,<1",
    "pre-commit>=4.0,<5",
    "types-PyYAML>=6.0,<7",
]

[project.scripts]
chaosengine = "errorworks.engine.cli:main"
chaosllm = "errorworks.llm.cli:main"
chaosllm-mcp = "errorworks.llm.cli:mcp_main_entry"
chaosweb = "errorworks.web.cli:main"

[project.urls]
Repository = "https://github.com/johnm-dta/errorworks"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/errorworks"]

[tool.hatch.build.targets.sdist]
exclude = [
    ".claude/",
    "CLAUDE.md",
    "docs/plans/",
    "docs/arch-analysis-*",
    "uv.lock",
]

# === Ruff Configuration ===
[tool.ruff]
target-version = "py312"
line-length = 140
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "W",      # pycodestyle warnings
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T20",    # flake8-print
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM108", # ternary operator - prefer explicit if/else
]

[tool.ruff.lint.isort]
known-first-party = ["errorworks"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["typer.Option", "typer.Argument"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["T20"]

# === Mypy Configuration ===
[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
show_error_codes = true
plugins = ["pydantic.mypy"]
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "hypothesis.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_decorators = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false

# === Pytest Configuration ===
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow",
    "stress: marks tests as stress tests",
    "chaosllm: Configure ChaosLLM server for the test",
    "chaosweb: Configure ChaosWeb server for the test",
    "integration: marks tests as integration tests",
]

# === Coverage Configuration ===
[tool.coverage.run]
source = ["src/errorworks"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
]
